+++ control-plane/aegisops_control_plane/__init__.py
@@ -20,7 +20,8 @@ from .models import (
     ReconciliationRecord,
     RecommendationRecord,
 )
-from .operations import RestoreReadinessService, RuntimeBoundaryService
+from .restore_readiness import RestoreReadinessService
+from .runtime_boundary import RuntimeBoundaryService
 from .service import (
     AegisOpsControlPlaneService,
     AnalystAssistantContextSnapshot,
